Move CodeSnippetDisplay into CodeSnippetWidget
